% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{adductAnnotation}
\alias{adductAnnotation}
\title{adductAnnotation}
\usage{
adductAnnotation(rMSIprocPeakMatrix, params, isotopicPatterns)
}
\description{
Given the monoisotopic ions found by the isotopes test, founds the possible adduct pairs considering the elements in the adductDataFrame.
}
