% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{annotationOutputFormat}
\alias{annotationOutputFormat}
\title{annotationOutputFormat}
\usage{
annotationOutputFormat(rMSIprocPeakMatrix, params, isotopeObj, adductObj)
}
\description{
This function gives format to the output
}
