% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLBinAppendIntensity}
\alias{CimzMLBinAppendIntensity}
\title{CimzMLBinAppendIntensity.
This function appends a new mass axis to a given ibd file.
The last added offset is returned.}
\usage{
CimzMLBinAppendIntensity(ibdFname, int_dataTypeString, intNew)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{int_dataTypeString:}{String to specify the data format used to encode m/z values.}

\item{intNew:}{The mass axis to append.}
}
\description{
CimzMLBinAppendIntensity.
This function appends a new mass axis to a given ibd file.
The last added offset is returned.
}
