% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLBinCreateNewIBD}
\alias{CimzMLBinCreateNewIBD}
\title{CimzMLBinCreateNewIBD.
This function creates a new ibd file with the provided uuid}
\usage{
CimzMLBinCreateNewIBD(ibdFname, str_uuid)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{uuid:}{16 bytes long UUID.}
}
\description{
CimzMLBinCreateNewIBD.
This function creates a new ibd file with the provided uuid
}
