% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkplotwidget.R
\name{createPlotWidget}
\alias{createPlotWidget}
\title{createPlotWidget.}
\usage{
createPlotWidget(
  parent,
  redraw_function,
  MouseSelection_callback = NULL,
  MouseWheel_callback = NULL,
  MouseHover_callback = NULL,
  initial_width = 480,
  initial_height = 480
)
}
\arguments{
\item{parent}{the gwidgets parent.}

\item{redraw_function}{a function to call to generate the plot.}

\item{MouseSelection_callback}{a callback function for the mouse click events.}

\item{MouseWheel_callback}{a callback function for the mouse wheel events.}

\item{MouseHover_callback}{a callback function for the mouse move events.}

\item{initial_width}{inital plot width in pixels.}

\item{initial_height}{inital plot height in pixels.}
}
\value{
a plotting widget
}
\description{
creates a plotting widget to use with gWidgets2tcltk and allows direct mouse points selection.
}
