% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMassImageByPeak.R
\name{.saveGgraphicsPlot2PngFile}
\alias{.saveGgraphicsPlot2PngFile}
\title{Save a gWidgets2::ggraphic widget to a png file.
A file dialog is presented to allow the user selecting where png file will be placed.}
\usage{
.saveGgraphicsPlot2PngFile(evt, ...)
}
\arguments{
\item{evt}{the evt$action must be a pointer to gWidgets2::ggraphic to save.}
}
\description{
Save a gWidgets2::ggraphic widget to a png file.
A file dialog is presented to allow the user selecting where png file will be placed.
}
