% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkhelpers.R
\name{gbutton_icon}
\alias{gbutton_icon}
\title{gbutton_icon.}
\usage{
gbutton_icon(image_file, container, handler)
}
\arguments{
\item{image_file}{}

\item{container}{}

\item{handler}{}
}
\value{
a gbutton widget
}
\description{
button with an image as icon.
}
