% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkhelpers.R
\name{gcombobox_rMSI}
\alias{gcombobox_rMSI}
\title{gcombobox_rMSI.}
\usage{
gcombobox_rMSI(items, selected = 1, container = NULL, handler = NULL)
}
\value{
a gcombobox widget
}
\description{
Build on top of gwidgets2tcltk implementation to properly set the widget width.
}
