% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkhelpers.R
\name{getValue_coloredCheckBox}
\alias{getValue_coloredCheckBox}
\title{getValue_coloredCheckBox.}
\usage{
getValue_coloredCheckBox(widget)
}
\value{
checkboz current value
}
\description{
get the value of a tcltk checkbox as boolean.
}
