% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{isotopeAnnotation}
\alias{isotopeAnnotation}
\title{isotopeAnnotation}
\usage{
isotopeAnnotation(rMSIprocPeakMatrix, params)
}
\description{
Finds and evaluate isotope candidates for each ion mass.
}
