% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadStorePeakMat.R
\name{LoadPeakMatrix}
\alias{LoadPeakMatrix}
\title{LoadPeakMatrix.}
\usage{
LoadPeakMatrix(data_path)
}
\arguments{
\item{data_path}{full path to .pkmat file where data is stored.}
}
\value{
a rMSI2 peak list object of the rMSIprocPeakMatrix class.
}
\description{
Loads a binned peaks matrix from HDD.
}
