% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procParams.R
\name{LoadProcParams}
\alias{LoadProcParams}
\title{LoadProcParams.}
\usage{
LoadProcParams(filename)
}
\arguments{
\item{filename}{full path to the .params file where processing paramers are stored.}
}
\value{
a rMSI2 processing parameters object.
}
\description{
Loads processing parameters from the HDD.
}
