% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkplotwidget.R
\name{mapDeviceCoords2PlotCoords}
\alias{mapDeviceCoords2PlotCoords}
\title{mapDeviceCoords2PlotCoords.}
\usage{
mapDeviceCoords2PlotCoords(x, y, plotting_data)
}
\arguments{
\item{x}{pixel X coordinate in the plotting device.}

\item{y}{pixel Y coordinate in the plotting device.}

\item{plotting_data}{a list object with the plot coords information}
}
\value{
a vector with X,Y coordinates in the data plot space.
}
\description{
Obtain the plot coordinates from pixel coordinates in the plotting device.
}
