% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overwriteIbdUUid}
\alias{overwriteIbdUUid}
\title{Method to overwrite the UUID of an imzML ibd file.
testingimzMLBinRead}
\usage{
overwriteIbdUUid(ibdFname, newUUID)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{newUUID:}{the new uuid as a string.}
}
\description{
Method to overwrite the UUID of an imzML ibd file.
testingimzMLBinRead
}
