% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnnotation.R
\name{peakAnnotation}
\alias{peakAnnotation}
\title{peakAnnotation}
\usage{
peakAnnotation(rMSIprocPeakMatrix, params)
}
\arguments{
\item{rMSIprocPeakMatrix}{An rMSIprocPeakMatrix object.}

\item{params}{Processing parameters object of the ProcParams class.}
}
\description{
Peak annotation tool based on the rMSIannotation algorithm Searches for carbon-based isotopic patterns in the peak matrix and evaluates them using morphology, intensity and mass error criteria. Later,
the algorithm searches adduct groups using the isotopic patterns.
}
\details{
Complete information on rMSIannotation can be found on the paper:
rMSIannotation: A peak annotation tool for mass spectrometry imaging based on the analysis of isotopic intensity ratios
Lluc Sementé, Gerard Baquer, María García-Altares, Xavier Correig-Blanchar, Pere Ràfols.

DOI: https://doi.org/10.1016/j.aca.2021.338669
}
\examples{
 
#First you need to load a peak matrix
pks <- rMSI2::LoadPeakMatrix("path/to/matrix.pkmat")

#Generate the results with default parameters
params <- rMSI2::ProcessingParameters() # Creates a default rMSI2 ProcParams object.
annotations <- rMSIproc::peakAnnotation(pks, params)

#Or with custom parameters
params <- rMSI2::rMSI2::LoadProcParams("path/to/parameters.params")
annotations <- rMSIproc::peakAnnotation(pks, params)

}
