% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiHelpers.R
\name{redrawPlotWidget}
\alias{redrawPlotWidget}
\title{redrawPlotWidget.
redraws the contents of a plot widget}
\usage{
redrawPlotWidget(tkplotwidget)
}
\arguments{
\item{tkplotwidget}{the plot widget to redraw.}
}
\description{
redrawPlotWidget.
redraws the contents of a plot widget
}
