% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librMSIdata.R
\name{ROIAverageSpectraByIds}
\alias{ROIAverageSpectraByIds}
\title{ROIAverageSpectraByIds.}
\usage{
ROIAverageSpectraByIds(img, Ids, maxMemMB = 200)
}
\arguments{
\item{img}{an rMSI object.}

\item{Ids}{Identifiers of spectra to use for average calculation.}

\item{maxMemMB}{maximum memory usage during spectral buffering in MB.}
}
\value{
the ROI average spectrum.
}
\description{
Calculates the average spectrum within each roi specified by a vector of pixel ID's
}
