% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkplotwidget.R
\name{setPlotWidgetDataRanges}
\alias{setPlotWidgetDataRanges}
\title{setPlotWidgetDataRanges.}
\usage{
setPlotWidgetDataRanges(
  tkplotwidget,
  data.range,
  plot.range,
  imgWin_width,
  imgWin_height
)
}
\arguments{
\item{tkplotwidget}{a tkrplot widget.}

\item{data.range}{data range obtanied with par().}

\item{plot.range}{plot range from par().}

\item{imgWin_width}{widget horizionalt size.}

\item{imgWin_height}{widget vertical size.}
}
\value{
nothing, the environment is modified directely.
}
\description{
method to set tkrplot widget plotting data ranges.
}
