% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procParams.R
\name{StoreProcParams}
\alias{StoreProcParams}
\title{StoreProcParams.}
\usage{
StoreProcParams(filename, params)
}
\arguments{
\item{filename}{full path to the .params file where processing paramers are stored.}

\item{params}{processing parameters object of the ProcParams class.}
}
\description{
Stores all processing parameters to HDD.
Parameters arestored compressed using RData format with .params extension.
}
