% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_additional_graphics.R
\name{theme_black}
\alias{theme_black}
\title{theme_black.
custom ggplot theme to display ion map images with a black background}
\usage{
theme_black(base_size = 12, base_family = "")
}
\description{
theme_black.
custom ggplot theme to display ion map images with a black background
}
