% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkhelpers.R
\name{coloredCheckBox}
\alias{coloredCheckBox}
\title{coloredCheckBox.}
\usage{
coloredCheckBox(
  text,
  checked = F,
  handler = NULL,
  container,
  foreground = NULL,
  background = NULL,
  bold = F
)
}
\arguments{
\item{text}{}

\item{checked}{}

\item{handler}{}

\item{container}{}

\item{foreground}{}

\item{background}{}

\item{bold}{}
}
\description{
A speciall checkbox using only tcltk to set colors and font types.
}
