% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiHelpers.R
\name{createPlotWidget}
\alias{createPlotWidget}
\title{createPlotWidget.
wrapper function to create tkrplots widgets integrated with gwidgets2}
\usage{
createPlotWidget(parent, redraw_function)
}
\arguments{
\item{parent}{gWidgets2 parent widget.}

\item{redraw_function}{function to draw the plot.}
}
\value{
a plot widget.
}
\description{
createPlotWidget.
wrapper function to create tkrplots widgets integrated with gwidgets2
}
