% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkhelpers.R
\name{gspinbutton_rMSI}
\alias{gspinbutton_rMSI}
\title{gspinbutton_rMSI.}
\usage{
gspinbutton_rMSI(
  from = 0,
  to = 10,
  by = 1,
  value = from,
  digits = 0,
  container = NULL,
  handler = NULL
)
}
\arguments{
\item{from}{}

\item{to}{}

\item{container}{}

\item{handler}{}
}
\value{
a gcombobox widget
}
\description{
Build on top of gwidgets2tcltk implementation to properly set the widget width.
}
