% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltkplotwidget.R
\name{redrawPlotWidget}
\alias{redrawPlotWidget}
\title{redrawPlotWidget.}
\usage{
redrawPlotWidget(tkplotwidget)
}
\arguments{
\item{tkplotwidget}{the widget to redraw.}
}
\description{
method to trigger the redraw of a tkrplot widget.
}
